Add module measurement capability for Linux 5.10
The trick was twofold, and represents gaps in my understanding. First, I had sometimes to invoke a translation table walk (directly) instead of using the more abstract translation function that decides over a table walk vs other types of address translation. That abstract function was not choosing the right methods of translation sometimes; it tried to treat kernel-module addresses as kernel addresses instead of user addresses, avoiding the table walks they required. The second part of the trick was discovering that the System.map addresses had an offset of 0x80000. In other words, the kernel is apparently not at the beginning of its address space. I wonder what those 0x80 pages are used for.