Add plaintext tables concerning the task_struct
The table in this commit is intended to represent the task_struct as it exists in this particular compilation-strategy for the 6.1 linux kernel. It remains to show the model is correct. The strategy used to build this table is as follows. The task_struct source in linux/sched.h was parsed for every precompiler directive that would effect the fields and layout of the struct. Every kernel variable in every precompiler conditional was dereferenced by using a kernel module. Those variables were then used with `gcc -E -P -nostdinc task.h` which resulted in a model I hope closely approximates the task_struct in memory. Then every type from every field was transformed into a size-in-bytes using a kernel module. Finally, the ultimate size of the task_struct was reported using a kernel module.