Partially Inspect the Process Tree
Our task-tree-crawling procedure is defunct. When we try to crawl a user-task, we are frequently failing to either (find the correct sibling pointer || not recognizing a terminal pointer of some sort). The crawl is bad and it's not immediately clear why. However, in this commit, it has been made easy to visualize the bytes of each task_struct crawled. More investigation is necessary.