measure VMAs via maple_tree traversal
A big commit, in many ways. In prior versions of linux, a task_struct's mm_struct had a simple reference to a circularly linked list of VMAs associated with the task. As of July 2022 (and here in linux 6.1.y), the VMAs are now stored in maple trees. The maple tree is a complex data structure, but it's not a bad target for introspection because it stands mostly alone and contains its own dump procedures. With some clever type finagling and pointer casting, and with some cherry-picked necessary inclusions, and with some intimate understanding of memory management, it was quite easy to write a new measurement over the VMAs.