Implement tests over signal and dataport communications.
This implements tests over every possible signal and dataport communication between the 3 camkes components. Of interest is that the VM done_wait signal requires its connection to be a seL4_GlobalAsynch in order for the consume_badge to be generated. Now, the wait signal is being consumed and dropped *not* by the client cakeml-am, but said client executes right after the signal consumption.