Fix handling of module_layout for linux 6.1
The module_layout offsets were hard-coded for an earlier version of linux. Additionally, this has uncovered a problem. The read-only data of the modules is not the same between boots due to some position-independent code strategy. Sometimes when the assembly instruction is "adrp," there is a virtual address as an operand that changes between boots. However, it is always the same relative to the base address of that kernel module. It's a complex problem, but that difference can be measured and accounted for.