Lift the am-cakeml's apps into the configure phase
The goal is this: build the am-cakeml's manifest-generator and -compiler during the cmake configure phase for this project. As well, we want to use the results of those am-cakeml apps. The generator will take a copland term to produce a manifest. The compiler will take the manifest to produce a server executable. Currently, the compiler is being built but not used.