Update formal manifests for both linux AM and CAmkES AM
Both AMs need to be armed with the mutual understanding of their aliases. From the perspective of both AMs, the CAmkES AM is Place 1, and the Linux AM is Place 0. Furthermore, Appraisal ASPs are denoted by where the measurement was conducted; because all measurements happen at Place 1, the appraisal ASPs should reflect that.