Greatly expand read-only data measurements
We now take a comprehensive approach over the kernel symbol table
specified in System.map. We add measurements over about 2600 more pages
in kernel memory; that's about 45% of all pages described in the kernel
table. This brings coverage up to about 61%. Most of what remains is
volatile data or data that will be unmapped after init.